Moved extra to separate file
